Refactoring / error handling
