+++ functions/assemblyscript/bytes.go
@@ -6,6 +6,7 @@ package assemblyscript
 
 import (
 	"context"
+	"errors"
 	"fmt"
 
 	wasm "github.com/tetratelabs/wazero/api"
@@ -19,7 +20,7 @@ func readBytes(mem wasm.Memory, offset uint32) ([]byte, error) {
 	// Read the length.
 	len, ok := mem.ReadUint32Le(offset - 4)
 	if !ok {
-		return nil, fmt.Errorf("failed to read buffer length")
+		return nil, errors.New("failed to read buffer length")
 	}
 
 	// Handle empty buffers.
@@ -30,12 +31,28 @@ func readBytes(mem wasm.Memory, offset uint32) ([]byte, error) {
 	// Now read the data into the buffer.
 	buf, ok := mem.Read(offset, len)
 	if !ok {
-		return nil, fmt.Errorf("failed to read buffer data from WASM memory")
+		return nil, errors.New("failed to read buffer data from WASM memory")
 	}
 
 	return buf, nil
 }
 
-func writeBytes(ctx context.Context, mod wasm.Module, bytes []byte) uint32 {
-	return writeObject(ctx, mod, bytes, 1)
+func writeBytes(ctx context.Context, mod wasm.Module, bytes []byte) (uint32, error) {
+	const classId = 1 // The fixed class id for an ArrayBuffer in AssemblyScript.
+	return writeRawBytes(ctx, mod, bytes, classId)
+}
+
+func writeRawBytes(ctx context.Context, mod wasm.Module, bytes []byte, classId uint32) (uint32, error) {
+	size := len(bytes)
+	offset, err := allocateWasmMemory(ctx, mod, size, classId)
+	if err != nil {
+		return 0, err
+	}
+
+	ok := mod.Memory().Write(offset, bytes)
+	if !ok {
+		return 0, fmt.Errorf("failed to write to WASM memory (%d bytes, class id %d)", size, classId)
+	}
+
+	return offset, nil
 }
