+++ functions/assemblyscript/dates.go
@@ -6,7 +6,6 @@ package assemblyscript
 
 import (
 	"context"
-	"encoding/binary"
 	"fmt"
 	"time"
 
@@ -33,12 +32,21 @@ func writeDate(ctx context.Context, mod wasm.Module, t time.Time) (uint32, error
 		}
 	}
 
-	t = t.UTC()
-	bytes := make([]byte, def.Size)
-	binary.LittleEndian.PutUint32(bytes, uint32(t.Year()))
-	binary.LittleEndian.PutUint32(bytes[4:], uint32(t.Month()))
-	binary.LittleEndian.PutUint32(bytes[8:], uint32(t.Day()))
-	binary.LittleEndian.PutUint64(bytes[16:], uint64(t.UnixMilli()))
+	offset, err := allocateWasmMemory(ctx, mod, int(def.Size), def.Id)
+	if err != nil {
+		return 0, err
+	}
 
-	return writeObject(ctx, mod, bytes, def.Id), nil
+	t = t.UTC()
+	mem := mod.Memory()
+	ok1 := mem.WriteUint32Le(offset, uint32(t.Year()))
+	ok2 := mem.WriteUint32Le(offset+4, uint32(t.Month()))
+	ok3 := mem.WriteUint32Le(offset+8, uint32(t.Day()))
+	ok4 := mem.WriteUint64Le(offset+16, uint64(t.UnixMilli()))
+
+	if !(ok1 && ok2 && ok3 && ok4) {
+		return 0, fmt.Errorf("failed to write Date object to WASM memory")
+	}
+
+	return offset, nil
 }
