+++ functions/assemblyscript/objects.go
@@ -40,18 +40,3 @@ func readObject(ctx context.Context, mem wasm.Memory, typ plugins.TypeInfo, offs
 
 	return readClass(ctx, mem, def, offset)
 }
-
-func writeObject(ctx context.Context, mod wasm.Module, bytes []byte, classId uint32) uint32 {
-	offset := allocateWasmMemory(ctx, mod, len(bytes), classId)
-	mod.Memory().Write(offset, bytes)
-	return offset
-}
-
-func allocateWasmMemory(ctx context.Context, mod wasm.Module, len int, classId uint32) uint32 {
-	// Allocate memory within the AssemblyScript module.
-	// This uses the `__new` function exported by the AssemblyScript runtime, so it will be garbage collected.
-	// See https://www.assemblyscript.org/runtime.html#interface
-	newFn := mod.ExportedFunction("__new")
-	res, _ := newFn.Call(ctx, uint64(len), uint64(classId))
-	return uint32(res[0])
-}
